youtube_search/requirements.txt
Django==1.8
django-cors-headers==1.2.2
djangorestframework==3.5.1
google-api-python-client==1.5.4
httplib2==0.9.2
MySQL-python==1.2.5
oauth2client==4.0.0
pkg-resources==0.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
